+++ C/helpers.h
@@ -10,8 +10,8 @@ void Cross3d(double a[3], double b[3], double c[3]);
 double Determinant(const int n, double a[n][n]);
 void InvertDirect(const int n, double A[n][n], double B[n][n]);
 double DotProduct(const int n, double a[n], double b[n]);
-void VecScale_(const int alpha, const int n, double v[n]);
-void MatScale_(const int alpha, const int m, const int n, double A[m][n]);
+void VecScale_(const double alpha, const int n, double v[n]);
+void MatScale_(const double alpha, const int m, const int n, double A[m][n]);
 void MatMult_(const int m, const int n, double A[m][n], double x[n], double b[m]);
 void MatMultTranspose_(const int m, const int n, double A[m][n], double x[m], double b[n]);
 void MatMatMult_(const int m, const int k, const int n, \
